[FIX] website_sale: show correct price of product not included in a pricelist discount Steps to reproduce: 1) website config: show strikethrough price (price comparison), show tax included price, discounts, pricelists(with advanced price rules). 2) create a pricelist with "show price and discount to customer", and any rule like 50% off any category (doesn't really matter). 3) apply the pricelist to the website
Issue:
Notice that some strikethrough price is applied wrongly - the crossed amount is the same as the written amount

Cause:
Strike is applied if base price not equal to price_reduce, which is actually calculated using taxes prior.
it should not be the case for the tax-included product ( or base price should also have tax applied).

opw-3140616