[FIX] web: fix default_export_compatible fields in Export dialog
This commit fixes the exported fields list used by the export dialog. Since the
list was deducted from the column titles, it was not possible to get the complete
attributes required to filter them. Now an extra-RPC is executed before the
dialog opens to fetch the correct list of exportable fields, with the correct
attributes.

It is now possible to filter fields correctly, and improve the reliability of the
export feature, since the filtering is no longer deducted from the JS, and uniquely
filtered manually as it was done in (1) for a fix.

(1): https://github.com/odoo/odoo/pull/111808

task-3203958