[IMP] core: create auto translate base fields
Before this commit:
Export/Import
Exporting a model_terms translated field will return the value of the user
language. Importing the modified model_terms translated field value will call
'write' api in the user language.
1. slow: most imported values are not changed after exported, and they still
have to SQL query all translations from database
2. lose translations: if the user only want to translate some terms for the
exported value, after import, all translations for the term will be lost. It is
because the `write` api for model_terms is not designed to be used for
translating

Views
Most views directly show model_terms translated fields to the user
Intuitively, users want to modify everything directly in the form view. If they
translate a term in the form view. The form view will call `write` api. All
translations for the term will be lost.

After this commit:
Export/Import
We can suggest the user don't export model_terms translated field and ask them
to export the translate_base field instead. Also for Views can display the
translate_base field if it is not readonly.
Since the translate_base field is always in en_US which is the source, changing
any term means changing the content of the term. And all translations for the
original term should be dropped.