[IMP] web_editor: add knowledge insert view embed requirement
The requirement is to store and make available the information that the
current user connected to a collaborative peer and its content was reset. There
is no way to know if such a connection will happen. The use case is inserting
content in the html_field programatically when a record is opened. At that time
the peerToPeer connection is not established, so the content should be inserted.
Then, if a collaborative session is started with someone already editing the
html_field, the content is reset, and at that point, the inserted element is
removed, so it should be inserted again.

This commit introduces a hook that can be used to execute some specific action
if the content of the editor is reset through `historyResetFromSteps` during
collaboration.

It also adds a test that validates the current behavior of the 'insert' command,
which unpacks the nodes to be inserted under certain conditions. This test can
be removed when an alternative to the command is implemented (or the command
stops unpacking content).

Task-2821374