[FIX] hr_expense: foreign currency is optional
By default expenses are now recorded in the company currency.
There is a company setting to re-enable to input expenses in
different currencies.

This shouldn't be the default behavior, as this feature is a
currency converted on the date of the expense, but does not yet
offer the possibility to reimburse expenses in foreign currency.

task-id: 3191597