[FIX] core: apply many2many commands to all corecords
The idea is to avoid applying the commands to active corecords only, and
apply them on all corecords instead.  The current behavior is confusing
and incorrect.  For instance, when "active_test" is not False, applying
Command.unlink(id) does not actually remove the given corecord if it is
inactive.  Moreover, the final result in cache is actually incorrect,
since inactive corecords have been silently discarded.