[IMP] web,crm,project: add an option to hide the "Add a Property" button
Purpose
=======
When this option is enabled, the button to create new properties
is invisible. We create a client action instead, and when the user
execute client action for the first time, the button become visible
until the user refresh the page.

This is useful because most of the time, the user will create the
properties and then won't change them. So this button use space
unnecessary in the form view (for the most of the use cases).

Task-3188915