[IMP] core: compare structure for model_terms translated fields
before this commit:
if a user tries to translate a model_terms translated field from the FORM view
with non-en_US context, all translations for user modified terms will be lost.

after this commit:
when the ORM detects the HTML/XML structure of the model_terms is not changed,
the diff of the terms will be treated as translation change rather than content
change.

Tips:
1. the current overhead is one extra `translate` per write per record
can be optimized, but may reduce the readability

2. one get_trans_terms in enterprise has to be modified