[FIX] base: fix the computed field that depends on a properties field
Bug
===
In Knowledge, we have a computed field, that depends on the properties
field. Because the cache of the parent is not updated in the onchange
method, in `convert_to_record`, when we get the definition, we read
the old value.

Because of that, the computed field is never triggered.

Task-3213639