[IMP] delivery: correct value in commercial invoices
Currently, For different shipping methods Commercial Invoice is generated,
in that unit value for each product is not accurate and subtotal of commercial
invoice is not equal to subtotal of all the products in the product line.

So in this commit, configure the calculation of sale_price which is used in
commercial invoice.