[IMP] web_editor: use a mutex to apply external steps
A Knowledge Behavior is a Component which is rendered asynchronously. This
rendering is done outside the editor in order to prevent the rendering mutations
from being registered in a step without preventing any step or mutation from
being registered at all (we don't want to split calls to observerUnactive or
historyPauseSteps from their counterpart between asynchronous processes, since
we would have to completely lock the editor during that time).

Nevertheless, some of the nodes in a Knowledge Behavior are collaborative (they
will share the same oids with the same nodes for each collaborator). While the
Behavior which they are a part of is being rendered outside of the editor, we
don't want to receive an external mutation from a collaborator related to those
oids, since they are not back yet in the editor and are thus unreachable.

Therefore, this commit introduces a mutex for collaborative interactions so
that async processes can finish their job before the next batch of collaborative
steps is applied.

This implies that a number of existing methods become asynchronous:
- onExternalHistorySteps
- _historyAddExternalStep
- onHistoryMissingParentSteps
- historyResetFromSteps
- (mergeClientsSteps) => test method

Side effect:
Since this commit makes use of the `concurrency` asset from the `web` module in
`OdooEditor.js`, in order to make tests work, they are converted to
`odoo-module` and a new bundle is created for the `mocha` tests of the editor:
`web_editor.mocha_tests`.

Task-2821374