[IMP] web_editor: test unpacking of the insert command
Currently, the 'insert' command unpacks nodes from a lone paragraph, or from
the first and last node if there are multiple nodes.

In knowledge, to append an embedded view to the editable, there is a need to
insert two sibling nodes at once (without unpacking them). With the current
behavior of 'insert', it is possible to do so by inserting a paragraph
containing the 2 nodes that need to be inserted. The result will be that the
'insert' command unpacks the parent paragraph and correctly insert the 2 nodes.

This commit adds a test which ensure that behavior is properly maintained, and
it can be removed when/if it is updated in the future. (Don't forget to update
the code in `wysiwyg.js:appendBehaviorBlueprint` in Knowledge at that point)

Task-2821374