[FIX] web_editor: fix cursor position after font color change
This commit addresses two different issues that led to the incorrect
placement of the cursor after a font or background color was selected
in the colorpicker.

Issue # 1:

Before this commit, the `applyColor` command returned non-connected nodes
(nodes with no parent) in some not well defined situations, leading to a Warning
when `range.setStart` was called in `_processAndApplyColor` with the
non-connected node as argument.

This commit properly removes such non-connected nodes from the array returned
by `applyColor`.

Issue # 2:

Before this commit, the cursor was not correctly placed when the font color
was selected with a collapsed range (which is possible on Mobile and when the
SnippetsMenu sidebar is available).

This commit fixes it by normalizing the selection when the cursor must be
placed inside a `<font>` element containing a zero-width space only.