[FIX] TO DELETE (fix in 15.0) web_editor: history manipulations should be reversible
Impacted Versions:
15.0+

Mutations that `add` nodes should always register a serialization of those nodes
at the moment the mutation occured, because their children could be modified
while they are not observed by the editor in a later step. If the DOM instances
of those nodes is used instead, it would result in a different DOM state when
the mutations are reverted.

1. Example without using serialization:

Let A be a node in the DOM observed by a MutationObserver (MO).
Let B, C be nodes outside of the DOM (not observed).

Execute operations:
```
DOM operation 1: "Add C to A", mutations:
- M1: Add C to A

DOM operation 2: "Add C to B", mutations:
- M2: Remove C from A
- Add C to B // that mutation has not been observed as B is not inside the DOM

DOM operation 3: "Add B to the DOM", mutations:
- M3: Add B to the DOM
```

Revert Steps:
```
- M3: Remove B from the DOM
- M2: Add C to A // consequently, C has been removed to B outside of the DOM
- M1: Remove C from A
```

Apply Steps:
```
- M1: Add C to A
- M2: Remove C from A
- M3: Add B to the DOM // B does not contain C
```

Conclusion:
We are missing a mutation from the operation 2 that was not registered, so when
we reapply the steps, C is missing from B.

2. How to reproduce in Odoo:
- Start an editor (without collaborative mode, since it already uses
  serialization). More specifically, `OdooEditor._collabClientId` should be
  falsy.
- The editable value should be the following, where `[]` is the cursor position:
```
<h1>text[]</h1>
```
- Type `Enter` (`oEnter` command)
- manually call `historyRevert` and `historyApply` on the last registered step
  of the editor

Current behavior:
- The result after those operations is:
```
<h1>text</h1>
<p></p>
```
The moved children are not properly re-added to their original parent.

Expected behavior:
```
<h1>text</h1>
<p><br></p>
```
The moved children are properly re-added to their original parent.

Fix:
Force the serialization of nodes in mutations used for history steps, and always
unserialize them instead of using cached nodes in `_idToNodeMap`, when the
mutation `add` nodes, in order to recover the configuration that they had when
the mutation was handled. This is not perfect because mutations are handled in
batches, and when we handle them, the nodes configuration is the result of the
entire batch, not the configuration when the mutation occured.

Since applying and reverting mutations can now cause the nodes instances to
change (since they are replaced by the unserialization of their state in a
mutation), restoring the selection during a `deleteRange` which could cause a
`historyRollback` has to be done using the `oid` of nodes instead of a
reference, in order to match their current instance in `_idToNodeMap`.

Task-3227060