[FIX] core: make env.lang a lazy property
This fixes a performance issue in a multit-threaded server when many
concurrent requests access env.lang (for accessing translated field, for
instance).  The implementation of env.lang relies on an ormcached method
to validate the current lang code, and concurrent accesses to the
ormcache seem to be more costly than expected.