[REF] core: `_read_group` for backend use
The `_read_group` was designed to be used by the web client to efficiently compute aggregations grouped by one or more fields. However, more and more developers have been using it from a backend perspective to make expensive computations more efficient. (Avoid doing the aggregation in Python.) Unfortunately, the API was designed for the web client, which added a lot of boilerplate when used in the backend.

`_read_group` was created to improve the performance of read_group for backend use, but didn't change the API and based the implementation on read_group itself.

Recreate `_read_group` from scratch with a new API to make it easier to use from the backend (see the method documentation). Also, split the method to make it easy to override and add custom behavior.