[WIP] mrp: display structure with use_model and with_search
The idea is to use the existing tools from owl and odoo framework js
The structure is define as
<Action>
  <WithSearch>
    <Main Component>
  </WithSearch>
</Action>

Since we use a client action and not a view the Action component is
use to define the fields and load their info from the server.
Also a default searchView is loaded at this step.

WithSearch do the link between the searchModel and the UseModel,
it will allow us to manipulate domain easily and reload the model
accordingly

Main Component call the UseModel with a type of list since the
action will work on multiple records. It also contain the layout