[FIX] web_editor: fix enforceWhitespace in pre
There was a special check for pre in enforceWhitespace. However,
this check was one condition too high, preventing the isVisibleStr
check below it to break the for loop when inside a P.

This resulted in the loop skipping the visible text node and then
finding an unrelated BR and duplicating it instead.

Task 3239097