[WIP] mrp: custom controlPanel + import searchView
The searchPanel is require for the client action. Since we already use
the searchModel, we do an extra rpc to search for the view with xmlId
`mrp_production_view_mrp_display`. This view contains a SearchPanel
inside so it's automaticaly compute. We could use the default search
view and set the searchPanel directly in the javascript but with the
xmlId it will be easy to override in order to add extra reasearch
fields.

If the view has been removed, we use the default one and display the
searchPanel in `mrp_display.js`.

We use a custom controlPanel to only handle 3 containers on a single line.