[WIP] mrp: define a model structure for client action model definition
The client action will use different models. Each models use differents
fields. It would be great to handle the records data at a single place

The generic useModel code is problematic because it takes the load
params from the props(from the searchModel). But since we use
different models, the seachView is only relative to the main and should
force its domain on other models (resulting in a traceback since they
don't share the same fields)