[IMP] account,point_of_sale: account tax repartition improvement
Before this commit:
===================
If the tax has a Tax Computation of Percentage of Price, then in the invoice
journal items lines for that tax calculations will  have the tax group and label
of the tax.
In the point of sale, there is no functionality to calculate tax by tax
repartition line.

After this commit:
===================
If the tax has a Tax Computation of Percentage of Price, then in the invoice
journal item lines for that tax calculations will have the tax group and label
of that tax repartition line if the tax repartition line has a value for the
invoice label and tax group.
Added a feature in the point of sale to calculate tax by tax repartition line
and show the calculation of tax in the point of sale receipt.