[FIX] l10n_in_edi : fixed label and documentation link in settings
l10n_in_edi : In Accounting settings the labels for Setup E-Invoice is
inconsistent with the other labels and the documentation link no
longer works. This commit fixes the label by using <span> instead
of <t> tag and updates the documentation link.

l10n_in_edi_ewaybill : In Accounting settings the labels for Setup E-Waybill
is inconsistent with the other labels and the documentation link no
longer works. This commit fixes the label by using <span> instead
of <t> tag and updates the documentation link.