[IMP] web: add supportedOptions documentation as field metadata
This commit adds a new supportedOptions attribute that can be added to any
field widget metadata object. This documentation can be used from other places
(Studio for example) by getting the field from the registry and then read
this key. This attribute can contain an array, detailing the list of available
options for the field widget.

Each option documented is an object and has the following attributes:
- name: name used in the options object from the node
- string: a label with a more explicit name than the option name
- type: the type of value that must be used as a value
- getChoices: can be used to generate some values for an option of type 'selection'
   where multiple values can be chosen as the value
- getValue: can be used to compute and return the current value of the option