[IMP] web, *: add supportedOptions documentation as field metadata
*: mail, mass_mailing, sms, web_editor

This commit adds a new supportedOptions attribute that can be added to any
field widget metadata object. This documentation can be used from other places
(Studio for example) by getting the field from the registry and then read
this key. This attribute can contain an array, detailing the list of available
options for the field widget.

By using an array, it can be ordered easily without having an object with keys
unordered alphabetically. Ordering the options can make sense, especially when
two options are tied to each other, it is easier to group them one after the
other. (eg: 'start_date' first, then 'end_date')

Each option documented is an object and has the following attributes:
- name: name used in the options object from the node
- label: a label with a more explicit name than the option name
- type: the type of value that must be used as a value
- choices: can be used to know available values for an option of type 'selection'
  where multiple values can be chosen as the value. This can be a function to
  generate the list of options, or an array containing the options.
- getValue: can be used to compute and return the current value of the option