[IMP] web: optimize many2one search
When searching for a many2one, the search is done with the "name_search"
method with the search term as part of the domain, with "ilike" operator.
However, when the search term is empty, it's useless to fetch with this
"ilike" operator, as it will return all the records.

So, when the search term is empty, we fetch all the records with the
"search_read" method, and we return them.

Task-id 3229757