[IMP] web, *: POC improve touch detection and behaviors
This commit is a Proof of concept, improving the touch experience in Odoo,
especially on hybrid devices (laptops with touchscreens or tablets with
mouse/keyboards). Without have the ability to detect the current input mode,
we can only verify the presence of 'touch' in the browser to adapt the UI/UX
accordingly. But those devices are exceptions to those mechanisms, on which
Odoo is currently not a great experience.

By having a way to have access to store the current input mode in a service,
it is now possible to get a better knowledge of the current 'isTouch' state
of the webclient.

The useAutofocus hook has also been improved. It makes no sense to autofocus
elements on touch devices, as the keyboard is really annoying when it pops for
no reason. Currently, it was only disabled on small devices, but this was
probably the case because of the annoyance of the touch keyboard on small screens.