[IMP] account: add "past" default filter on `account.move.line`
The on `account.move.line`, the default order is `date desc, move_name
desc, i`.
There is a corresponding index `account_move_line_date_name_id_idx`, on
which we perform an index scan.
Because there can be a lot of moves in draft in the future, the index
scan still needs to remove a lot of entries because of the default
filter (`state` is `posted`) doesn't apply for the vast majority of the
moves in the future.
Because there shouldn't be posted moves in the future, it is safe to add
a default filter on top of `Posted` to help the index scan and start
directly where needed.