[IMP] tests: Clear precommits on test cleanup
Precommit hooks would stock data until a call to ``cr.flush`` was made.
Notably, this happens when the ``assertRaises`` method is called.
Functions were applied on records already cleared from the cache.

This change adds a cleanup call for `TransactionCase` and as it keeps
the same cursor for all tests. Cursor precommits can now
be safely executed inside tests.

Task-2834304