[IMP] mail: compress old messages
Description:
For large companies, the `mail_message` model can get quite large.
One of the primary culprit is the `body` field, which is an HTML field.

Solution:
Create a cron (inactive by default) that convert old
`mail_message.body` to plaintext and removes quotes from it's content.
Removing styling and html tags + the quotes reduces the space used to
store the body of the messages.

perf-3209468