[MOV] *: all addons to better places (not farms upstate)
- move all regular business addons to odoo/addons
- move all test addons do `addons_test` at the repo root
- move all `hw`-related addons (IOT) to `addons_hw` at the repo root,
  as they often have particular concerns in terms of dependencies and
  deployments

Use a layout for the `addons_*` directories which is compatible with
PEP 420 and not having to use addons paths.

Q / TODO: is account_test a test addon? Its description hints that it
should be installed on production system in order to create and apply
accounting consistency rules, but it's not entirely clear.