[FIX] web_editor: fix "remove link" behavior
Weird stuff happen when a link is unlinked before the linkpopover is
fully hidden.

More specifically: the editor's observer is switched off when the
popover begins its hiding process, which is triggered upon clicking the
"unlink" button, and it's only switched back on when the hiding is
complete. Executing the 'unlink' command with the observer switched off
leads to bad history registration.

On top of that, according to Bootstrap's doc [1]:
"Popovers must be hidden before their corresponding elements have been
removed from the DOM."

Moreover, the recovery of the isolated link hack after a command has
been applied to a link makes little sense if that command was 'unlink'.

[1]: https://getbootstrap.com/docs/4.0/components/popovers/