[IMP] web_editor: update popover upon link changes
Before this commit, the link popover would only update it's content and
position when it was being shown. Once shown, its content and position
would be outdated (possibly covering the link's new label) if any change
happened to the link's label or href (note that a change in a link's label
can change its href).

This commit makes sure the popover stays always up-to-date.