[FIX] mrp(_workorder): backorder created as per operation type
Before this commit:

The option on the operation type to select the way in which backorders should be
generated (never, always, ask) does not work for the Manufacturing Operation
type.

After this commit:

The option on the operation type to select the way in which backorders should be
generated will work for the Manufacturing Operation type.

Task-2918994