[IMP] mrp: sub component reactivity
We pass the parent as a props because an action on child component
requires a check on the main record (e.g. it could be in the to_close
state and display the validation button).

Reloading everything is probably too much but an rpc check will be made
at the end to check that