[FIX] web_editor: persist URL-like text transformation to link
TARGET: 16.0
TODO: 	check if [1]'s fix is not broken
	write this case in a tour

Commit [1] separated the 2 concerns handled by _onURLInput: making
changes to the DOM and updating the UI, taking care that only the second
one is done at start. But this broke the following functionality:

- Have a URL-like text (ex: odoo.com), select it and click the link icon
on the sidebar.
- The URL is deduced from the link's label.
- Click any anywhere to close the sidebar.

Before [1], the link transformation would be kept. After it, it is not.
This happens because this is a particular case in which we want to apply
changes to the DOM at start time, was if a URL input event had happened.

This commit fixes it.

[1]: 54746012d37c457d4c2f32d419b03f21cd8718f9