[IMP] l10n_in_edi:  Before Test connection check GST number
In this commit:
=================
This commit adds functionality to check if the GST number is present before
testing the connection for the l10n_in_edi module. This check ensures that the
provided GST number is present, helping to prevent authentication errors that
can occur if the GST number is missing

The error message has also been improved to provide better feedback to the user
i
f the GST number is not present.

This change improves the reliability and user experience of the l10n_in_edi
module.